+++ gegelatilib/src/file/parametersParser.cpp
@@ -62,23 +62,24 @@ void File::ParametersParser::readConfigFile(const char* path, Json::Value& root)
 void File::ParametersParser::setAllParamsFrom(const Json::Value& root,
                                               Learn::LearningParameters& params)
 {
-    for (std::string& key : root.getMemberNames()) {
+    for (std::string const& key : root.getMemberNames()) {
         if (key == "mutation") {
             // we have a subtree of mutation : parameters like mutation.xxx.xxx
             for (std::string const& key2 : root[key].getMemberNames()) {
                 if (key2 == "tpg") {
                     // we're on a mutation.tpg.xxx parameter
-                    for (std::string& key3 : root[key][key2].getMemberNames()) {
-                        double value = root[key][key2][key3].asDouble();
+                    for (std::string const& key3 :
+                         root[key][key2].getMemberNames()) {
+                        Json::Value value = root[key][key2][key3];
                         setParameterFromString(params, key3, value);
                     }
                 }
                 else {
                     if (key2 == "prog") {
                         // we're on a mutation.prog.xxx parameter
-                        for (std::string& key3 :
+                        for (std::string const& key3 :
                              root[key][key2].getMemberNames()) {
-                            double value = root[key][key2][key3].asDouble();
+                            Json::Value value = root[key][key2][key3];
                             setParameterFromString(params, key3, value);
                         }
                     }
@@ -88,109 +89,114 @@ void File::ParametersParser::setAllParamsFrom(const Json::Value& root,
         }
         if (root[key].size() == 0) {
             // we have a parameter without subtree (as a leaf)
-            double value = root[key].asDouble();
+            Json::Value value = root[key];
             setParameterFromString(params, key, value);
         }
     }
 }
 
 void File::ParametersParser::setParameterFromString(
-    Learn::LearningParameters& params, std::string& key, double value)
+    Learn::LearningParameters& params, const std::string& param,
+    Json::Value const& value)
 {
-    if (key == "nbActions") {
-        params.mutation.tpg.nbActions = (size_t)value;
+    if (param == "nbActions") {
+        params.mutation.tpg.nbActions = (size_t)value.asUInt();
         return;
     }
-    if (key == "nbRoots") {
-        params.mutation.tpg.nbRoots = (size_t)value;
+    if (param == "nbRoots") {
+        params.mutation.tpg.nbRoots = (size_t)value.asUInt();
         return;
     }
-    if (key == "maxInitOutgoingEdges") {
-        params.mutation.tpg.maxInitOutgoingEdges = (size_t)value;
+    if (param == "maxInitOutgoingEdges") {
+        params.mutation.tpg.maxInitOutgoingEdges = (size_t)value.asUInt();
         return;
     }
-    if (key == "maxOutgoingEdges") {
-        params.mutation.tpg.maxOutgoingEdges = (size_t)value;
+    if (param == "maxOutgoingEdges") {
+        params.mutation.tpg.maxOutgoingEdges = (size_t)value.asUInt();
         return;
     }
-    if (key == "pEdgeDeletion") {
-        params.mutation.tpg.pEdgeDeletion = value;
+    if (param == "pEdgeDeletion") {
+        params.mutation.tpg.pEdgeDeletion = value.asDouble();
         return;
     }
-    if (key == "pEdgeAddition") {
-        params.mutation.tpg.pEdgeAddition = value;
+    if (param == "pEdgeAddition") {
+        params.mutation.tpg.pEdgeAddition = value.asDouble();
         return;
     }
-    if (key == "pProgramMutation") {
-        params.mutation.tpg.pProgramMutation = value;
+    if (param == "pProgramMutation") {
+        params.mutation.tpg.pProgramMutation = value.asDouble();
         return;
     }
-    if (key == "pEdgeDestinationChange") {
-        params.mutation.tpg.pEdgeDestinationChange = value;
+    if (param == "pEdgeDestinationChange") {
+        params.mutation.tpg.pEdgeDestinationChange = value.asDouble();
         return;
     }
-    if (key == "pEdgeDestinationIsAction") {
-        params.mutation.tpg.pEdgeDestinationIsAction = value;
+    if (param == "pEdgeDestinationIsAction") {
+        params.mutation.tpg.pEdgeDestinationIsAction = value.asDouble();
         return;
     }
-    if (key == "maxProgramSize") {
-        params.mutation.prog.maxProgramSize = (size_t)value;
+    if (param == "maxProgramSize") {
+        params.mutation.prog.maxProgramSize = (size_t)value.asUInt();
         return;
     }
-    if (key == "pDelete") {
-        params.mutation.prog.pDelete = value;
+    if (param == "pDelete") {
+        params.mutation.prog.pDelete = value.asDouble();
         return;
     }
-    if (key == "pAdd") {
-        params.mutation.prog.pAdd = value;
+    if (param == "pAdd") {
+        params.mutation.prog.pAdd = value.asDouble();
         return;
     }
-    if (key == "pMutate") {
-        params.mutation.prog.pMutate = value;
+    if (param == "pMutate") {
+        params.mutation.prog.pMutate = value.asDouble();
         return;
     }
-    if (key == "pSwap") {
-        params.mutation.prog.pSwap = value;
+    if (param == "pSwap") {
+        params.mutation.prog.pSwap = value.asDouble();
         return;
     }
-    if (key == "archiveSize") {
-        params.archiveSize = (size_t)value;
+    if (param == "archiveSize") {
+        params.archiveSize = (size_t)value.asUInt();
         return;
     }
-    if (key == "archivingProbability") {
-        params.archivingProbability = value;
+    if (param == "archivingProbability") {
+        params.archivingProbability = value.asDouble();
         return;
     }
-    if (key == "nbIterationsPerPolicyEvaluation") {
-        params.nbIterationsPerPolicyEvaluation = (uint64_t)value;
+    if (param == "nbIterationsPerPolicyEvaluation") {
+        params.nbIterationsPerPolicyEvaluation = value.asUInt64();
         return;
     }
-    if (key == "maxNbActionsPerEval") {
-        params.maxNbActionsPerEval = (uint64_t)value;
+    if (param == "maxNbActionsPerEval") {
+        params.maxNbActionsPerEval = value.asUInt64();
         return;
     }
-    if (key == "ratioDeletedRoots") {
-        params.ratioDeletedRoots = value;
+    if (param == "ratioDeletedRoots") {
+        params.ratioDeletedRoots = value.asDouble();
         return;
     }
-    if (key == "nbGenerations") {
-        params.nbGenerations = (uint64_t)value;
+    if (param == "nbGenerations") {
+        params.nbGenerations = value.asUInt64();
         return;
     }
-    if (key == "maxNbEvaluationPerPolicy") {
-        params.maxNbEvaluationPerPolicy = (size_t)value;
+    if (param == "maxNbEvaluationPerPolicy") {
+        params.maxNbEvaluationPerPolicy = (size_t)value.asUInt();
         return;
     }
-    if (key == "nbRegisters") {
-        params.nbRegisters = (size_t)value;
+    if (param == "nbRegisters") {
+        params.nbRegisters = (size_t)value.asUInt();
         return;
     }
-    if (key == "nbThreads") {
-        params.nbThreads = (size_t)value;
+    if (param == "nbThreads") {
+        params.nbThreads = (size_t)value.asUInt();
+        return;
+    }
+    if (param == "doValidation") {
+        params.doValidation = value.asBool();
         return;
     }
     // we didn't recognize the symbol
-    std::cerr << "Ignoring unknown parameter " << key << std::endl;
+    std::cerr << "Ignoring unknown parameter " << param << std::endl;
 }
 
 void File::ParametersParser::loadParametersFromJson(