+++ gegelatilib/src/log/LABasicLogger.cpp
@@ -3,27 +3,7 @@
 
 #include "log/LABasicLogger.h"
 
-void Log::LABasicLogger::logHeader()
-{
-    // fixing float precision
-    *this << std::setprecision(2) << std::fixed << std::left;
-    *this << std::setw(colWidth) << "Gen" << std::setw(colWidth) << "NbVert"
-          << std::setw(colWidth) << "Min" << std::setw(colWidth) << "Avg"
-          << std::setw(colWidth) << "Max" << std::setw(colWidth)
-          << "Duration(eval)" << std::setw(colWidth) << "Duration(decim)"
-          << std::setw(colWidth) << "Total_time" << std::endl;
-}
-
-void Log::LABasicLogger::logAfterPopulateTPG(uint64_t& generationNumber,
-                                             TPG::TPGGraph& tpg)
-{
-    *this << std::setw(colWidth) << generationNumber << std::setw(colWidth)
-          << tpg.getNbVertices();
-    // resets checkpoint to be able to show evaluation time
-    chronoFromNow();
-}
-
-void Log::LABasicLogger::logAfterEvaluate(
+void Log::LABasicLogger::logResults(
     std::multimap<std::shared_ptr<Learn::EvaluationResult>,
                   const TPG::TPGVertex*>& results)
 {
@@ -39,14 +19,63 @@ void Log::LABasicLogger::logAfterEvaluate(
                pair) -> double { return acc + pair.first->getResult(); });
     avg /= (double)results.size();
     *this << std::setw(colWidth) << min << std::setw(colWidth) << avg
-          << std::setw(colWidth) << max << std::setw(colWidth)
-          << getDurationFrom(*checkpoint);
-    // resets checkpoint to be able to show decimation time
+          << std::setw(colWidth) << max;
+}
+
+void Log::LABasicLogger::logHeader()
+{
+    // fixing float precision
+    *this << std::setprecision(2) << std::fixed << std::left;
+    *this << std::setw(colWidth) << "Gen" << std::setw(colWidth) << "NbVert"
+          << std::setw(colWidth) << "Min" << std::setw(colWidth) << "Avg"
+          << std::setw(colWidth) << "Max" << std::setw(colWidth)
+          << "Duration(eval)";
+    if (doValidation) {
+        *this << std::setw(colWidth) << "Duration(valid)";
+    }
+    *this << std::setw(colWidth) << "Total_time" << std::endl;
+}
+
+void Log::LABasicLogger::logAfterPopulateTPG(uint64_t& generationNumber,
+                                             TPG::TPGGraph& tpg)
+{
+    *this << std::setw(colWidth) << generationNumber << std::setw(colWidth)
+          << tpg.getNbVertices();
+    // resets checkpoint to be able to show evaluation time
     chronoFromNow();
 }
 
-void Log::LABasicLogger::logAfterDecimate(TPG::TPGGraph& tpg)
+void Log::LABasicLogger::logAfterEvaluate(
+    std::multimap<std::shared_ptr<Learn::EvaluationResult>,
+                  const TPG::TPGVertex*>& results)
 {
-    *this << std::setw(colWidth) << getDurationFrom(*checkpoint)
-          << std::setw(colWidth) << getDurationFrom(*start) << std::endl;
+    evalTime = getDurationFrom(*checkpoint);
+
+    // we only log results statistics if there is no validation
+    if (!doValidation) {
+        logResults(results);
+    }
+
+    // resets checkpoint to be able to show validation time if there is some
+    chronoFromNow();
+}
+
+void Log::LABasicLogger::logAfterValidate(
+    std::multimap<std::shared_ptr<Learn::EvaluationResult>,
+                  const TPG::TPGVertex*>& results)
+{
+    validTime = getDurationFrom(*checkpoint);
+
+    // being in this method means validation is active, and so we are sure we
+    // can log results
+    logResults(results);
+}
+
+void Log::LABasicLogger::logEndOfTraining()
+{
+    *this << std::setw(colWidth) << evalTime;
+    if (doValidation) {
+        *this << std::setw(colWidth) << validTime;
+    }
+    *this << std::setw(colWidth) << getDurationFrom(*start) << std::endl;
 }