+++ test/LABasicLoggerTest.cpp
@@ -49,7 +49,7 @@ class LABasicLoggerTest : public ::testing::Test
     Instructions::Set set;
     Environment* e = NULL;
     TPG::TPGGraph* tpg = NULL;
-    uint64_t generation = 0;
+    uint64_t generation = 1;
     std::multimap<std::shared_ptr<Learn::EvaluationResult>,
                   const TPG::TPGVertex*>* results =
         new std::multimap<std::shared_ptr<Learn::EvaluationResult>,
@@ -94,13 +94,22 @@ class LABasicLoggerTest : public ::testing::Test
     }
 };
 
-TEST_F(LABasicLoggerTest, ConstructorAndlogHeader)
+TEST_F(LABasicLoggerTest, Constructor)
 {
     ASSERT_NO_THROW(Log::LABasicLogger l);
     ASSERT_NO_THROW(Log::LABasicLogger l(std::cerr));
+}
 
+TEST_F(LABasicLoggerTest, logHeader)
+{
     std::stringstream strStr;
     Log::LABasicLogger l(strStr);
+
+    // basic header without validation
+    l.logHeader();
+    // we log a second header with validation column
+    l.doValidation = true;
+    l.logHeader();
 
     // now we will check the header logged correctly
     std::string s = strStr.str();
@@ -116,8 +125,8 @@ TEST_F(LABasicLoggerTest, ConstructorAndlogHeader)
     ASSERT_EQ("Avg", result[3]);
     ASSERT_EQ("Max", result[4]);
     ASSERT_EQ("Duration(eval)", result[5]);
-    ASSERT_EQ("Duration(decim)", result[6]);
-    ASSERT_EQ("Total_time", result[7]);
+    ASSERT_EQ("Total_time", result[6]);
+    ASSERT_EQ("Duration(valid)", result[13]);
 }
 
 TEST_F(LABasicLoggerTest, logAfterPopulateTPG)
@@ -133,8 +142,8 @@ TEST_F(LABasicLoggerTest, logAfterPopulateTPG)
     for (std::string s2; iss >> s2;)
         result.push_back(s2);
 
-    ASSERT_EQ("0", result[8]);
-    ASSERT_EQ("0", result[9]);
+    ASSERT_EQ("1", result[0]);
+    ASSERT_EQ("0", result[1]);
 }
 
 TEST_F(LABasicLoggerTest, logAfterEvaluate)
@@ -150,24 +159,17 @@ TEST_F(LABasicLoggerTest, logAfterEvaluate)
     for (std::string s2; iss >> s2;)
         result.push_back(s2);
 
-    ASSERT_DOUBLE_EQ(5.00, std::stod(result[8]));
-    ASSERT_DOUBLE_EQ(7.50, std::stod(result[9]));
-    ASSERT_DOUBLE_EQ(10.00, std::stod(result[10]));
-    ASSERT_TRUE(std::stod(result[11]) >= 0.0)
-        << "eval duration should be positive";
+    ASSERT_DOUBLE_EQ(5.00, std::stod(result[0]));
+    ASSERT_DOUBLE_EQ(7.50, std::stod(result[1]));
+    ASSERT_DOUBLE_EQ(10.00, std::stod(result[2]));
 }
 
-TEST_F(LABasicLoggerTest, logAfterDecimate)
+TEST_F(LABasicLoggerTest, logAfterValidate)
 {
     std::stringstream strStr;
     Log::LABasicLogger l(strStr);
 
-    // little sleep to delay the total_time value (while the "checkpoint" of the
-    // logger will be reset)
-    double timeToWaitMili = 10;
-    std::this_thread::sleep_for(std::chrono::milliseconds(10));
-
-    l.logAfterDecimate(*tpg);
+    l.logAfterValidate(*results);
     std::string s = strStr.str();
     // putting each element seperated by blanks in a tab
     std::vector<std::string> result;
@@ -175,17 +177,25 @@ TEST_F(LABasicLoggerTest, logAfterDecimate)
     for (std::string s2; iss >> s2;)
         result.push_back(s2);
 
-    double decimTime = std::stod(result[8]);
-    double totTime = std::stod(result[9]);
-    ASSERT_TRUE(decimTime >= 0.0) << "decimation duration should be positive";
-    ASSERT_TRUE(totTime >= 0.0) << "total elapsed time should be positive";
+    ASSERT_DOUBLE_EQ(5.00, std::stod(result[0]));
+    ASSERT_DOUBLE_EQ(7.50, std::stod(result[1]));
+    ASSERT_DOUBLE_EQ(10.00, std::stod(result[2]));
 }
 
-TEST_F(LABasicLoggerTest, ChronoFromNow)
+TEST_F(LABasicLoggerTest, logAfterDecimate)
 {
-    // to test chrono, we will wait, use chronoFromNow() which resets the
-    // "checkpoint" time and call logAfterDecimate() which shall print the
-    // duration from checkpoint and from start
+    std::stringstream strStr;
+    Log::LABasicLogger l(strStr);
+    ASSERT_NO_THROW(l.logAfterDecimate(*tpg));
+}
+
+TEST_F(LABasicLoggerTest, logEndOfTraining)
+{
+    // To test chrono, we will wait, use chronoFromNow() which resets the
+    // "checkpoint" time, then call logAfterEvaluate() which will register
+    // evalTime and call logEndOfTraining() which shall log the duration from
+    // checkpoint and from start.
+    // The total duration should be larger than the evalTime
 
     std::stringstream strStr;
     Log::LABasicLogger l(strStr);
@@ -198,8 +208,13 @@ TEST_F(LABasicLoggerTest, ChronoFromNow)
     // resets "checkpoint" so that the first displayed time shall be lower than
     // the second which is the time from start
     l.chronoFromNow();
+    l.doValidation = true; // to avoid logging eval statistics
+    l.logAfterEvaluate(*results);
+    l.logEndOfTraining();
+    // then, we can test the method when there is no validation
+    l.doValidation = false; // to avoid logging eval statistics
+    l.logEndOfTraining();
 
-    l.logAfterDecimate(*tpg);
     std::string s = strStr.str();
     // putting each element seperated by blanks in a tab
     std::vector<std::string> result;
@@ -207,10 +222,16 @@ TEST_F(LABasicLoggerTest, ChronoFromNow)
     for (std::string s2; iss >> s2;)
         result.push_back(s2);
 
-    double decimTime = std::stod(result[8]);
-    double totTime = std::stod(result[9]);
-    ASSERT_TRUE(totTime > decimTime)
+    double evalTime = std::stod(result[0]);
+    double validTime = std::stod(result[1]);
+    double totTime = std::stod(result[2]);
+    ASSERT_TRUE(evalTime >= 0) << "Eval duration should be positive";
+    ASSERT_TRUE(validTime >= 0) << "Valid duration should be positive";
+    ASSERT_TRUE(totTime > evalTime)
         << "Total time should be the largest duration !";
     ASSERT_TRUE(totTime >= timeToWaitMili / 1000)
         << "Total time should be larger than the time we waited !";
+
+    ASSERT_TRUE(result.size() == 5)
+        << "logEndOfTraining with and without valid should have 3+2=5 elements";
 }